# docs/swagger.yaml
basePath: /
definitions:
  model.Config:
    properties:
      labels:
        additionalProperties:
          type: string
        description: |-
          Labels are key-value pairs for labeling the configuration
          Required: true
        type: object
      name:
        description: |-
          Name of the configuration
          Required: true
        type: string
      params:
        additionalProperties:
          type: string
        description: |-
          Parameters are key-value pairs for configuration
          Required: true
        type: object
      version:
        description: |-
          Version of the configuration
          Required: true
        type: integer
    type: object
  model.ConfigGroup:
    properties:
      config:
        items:
          $ref: '#/definitions/model.Config'
        type: array
      name:
        type: string
      version:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for a configuration service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Configuration API
  version: "1.0"
paths:
  /configGroups:
    post:
      consumes:
      - application/json
      description: Adds a new configuration group
      parameters:
      - description: Configuration group to add
        in: body
        name: configGroup
        required: true
        schema:
          $ref: '#/definitions/model.ConfigGroup'
      produces:
      - application/json
      responses:
        "200":
          description: Configuration group added successfully
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Add a new configuration group
      tags:
      - configGroups
  /configGroups/{groupName}/{groupVersion}/{configName}/{configVersion}:
    delete:
      description: Removes a specific configuration from a specified group
      parameters:
      - description: Name of the configuration group
        in: path
        name: groupName
        required: true
        type: string
      - description: Version of the configuration group
        in: path
        name: groupVersion
        required: true
        type: integer
      - description: Name of the configuration to remove
        in: path
        name: configName
        required: true
        type: string
      - description: Version of the configuration to remove
        in: path
        name: configVersion
        required: true
        type: integer
      responses:
        "200":
          description: Configuration removed from group successfully
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Configuration or group not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Remove a configuration from a group
      tags:
      - configGroups
  /configGroups/{groupName}/{groupVersion}/{labels}:
    delete:
      description: Deletes all configurations within a specific group that match the
        specified labels
      parameters:
      - description: Name of the configuration group
        in: path
        name: groupName
        required: true
        type: string
      - description: Version of the configuration group
        in: path
        name: groupVersion
        required: true
        type: integer
      - description: Labels to filter the configurations for deletion
        in: path
        name: labels
        required: true
        type: string
      responses:
        "200":
          description: Configurations deleted successfully
          schema:
            type: string
        "400":
          description: Invalid group version
          schema:
            type: string
        "404":
          description: Configurations not found for deletion
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete configurations by labels
      tags:
      - configGroups
    get:
      description: Retrieves all configurations within a specific group that match
        the specified labels
      parameters:
      - description: Name of the configuration group
        in: path
        name: groupName
        required: true
        type: string
      - description: Version of the configuration group
        in: path
        name: groupVersion
        required: true
        type: integer
      - description: Labels to filter the configurations
        in: path
        name: labels
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of configurations
          schema:
            items:
              $ref: '#/definitions/model.Config'
            type: array
        "400":
          description: Invalid group version
          schema:
            type: string
        "404":
          description: Configurations not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Retrieve configurations by labels
      tags:
      - configGroups
  /configGroups/{name}/{version}:
    delete:
      description: Deletes a configuration group by name and version
      parameters:
      - description: Name of the configuration group
        in: path
        name: name
        required: true
        type: string
      - description: Version of the configuration group
        in: path
        name: version
        required: true
        type: integer
      responses:
        "200":
          description: Configuration group deleted successfully
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Configuration group not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete a configuration group
      tags:
      - configGroups
    get:
      description: Retrieves a configuration group by name and version
      parameters:
      - description: Name of the configuration group
        in: path
        name: name
        required: true
        type: string
      - description: Version of the configuration group
        in: path
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ConfigGroup'
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Configuration group not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get a configuration group
      tags:
      - configGroups
  /configGroups/{name}/{version}/configs:
    post:
      consumes:
      - application/json
      description: Adds a configuration to a specified group
      parameters:
      - description: Name of the configuration group
        in: path
        name: name
        required: true
        type: string
      - description: Version of the configuration group
        in: path
        name: version
        required: true
        type: integer
      - description: Configuration to add to the group
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/model.Config'
      produces:
      - application/json
      responses:
        "200":
          description: Configuration added to group successfully
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Configuration group not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Add a configuration to a group
      tags:
      - configGroups
  /configs:
    post:
      consumes:
      - application/json
      description: Adds a new configuration
      parameters:
      - description: Configuration to add
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/model.Config'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Config'
        "400":
          description: Invalid input
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Add a new configuration
      tags:
      - configs
  /configs/{name}/{version}:
    delete:
      description: Deletes a configuration by name and version
      parameters:
      - description: Name of the configuration
        in: path
        name: name
        required: true
        type: string
      - description: Version of the configuration
        in: path
        name: version
        required: true
        type: integer
      responses:
        "200":
          description: Successfully deleted configuration
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Configuration not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete a configuration
      tags:
      - configs
    get:
      description: Retrieves a configuration by name and version
      parameters:
      - description: Name of the configuration
        in: path
        name: name
        required: true
        type: string
      - description: Version of the configuration
        in: path
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Config'
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Configuration not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get a configuration
      tags:
      - configs
swagger: "2.0"